Add JAVA_EBUILDER variable to scripts to support command aliasing
Sometimes it would be helpful to alias the java-ebuilder command used in
the Makefiles, so a java-ebuilder JAR different from the one installed
to the system could be used.  For example, developers who wish to test
their changes to java-ebuilder might want to run with the JAR built from
the new source code.  This change would enable them to do this with
something like

    JAVA_EBUILDER="java -jar ../../../target/java-ebuilder-SNAPSHOT.jar" \
        make build

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>